requirements.txt
astroid==2.0.4
autopep8==1.4.2
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.6
requests==2.20.0
simplejson==3.16.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24
wrapt==1.10.11
